+++ packages/theme/src/composables/config/index.ts
@@ -107,6 +107,16 @@ export namespace Theme {
      * @default true
      */
     mobileMinify?: boolean
+    /**
+     * 指定按钮距离屏幕下边缘的距离
+     * @default 40
+     */
+    marginBottom?: number
+    /**
+     * 指定图标的大小
+     * @default 20
+     */
+    iconSize?: number
   }
   export interface GiscusConfig extends CommentCommonConfig {
     type: 'giscus'