+++ packages/vitepress-plugin-giscus/src/components/GiscusComment.vue
@@ -76,6 +76,12 @@ const labelText = computed(() => {
 const showCommentBtn = computed(() => {
   return commentConfig.value?.showCommentBtn ?? true
 })
+
+const defaultMarginBottom = 40
+const marginBottom = computed(() => `${commentConfig.value?.marginBottom ?? defaultMarginBottom}px`)
+
+const defaultIconSize = 20
+const iconSize = computed(() => commentConfig.value?.iconSize ?? defaultIconSize)
 </script>
 
 <template>
@@ -95,16 +101,16 @@ const showCommentBtn = computed(() => {
     />
     <div v-if="showCommentBtn && _docWidth" v-show="!commentIsVisible" class="comment-btn-wrapper" :class="{ hidden: commentIsVisible }">
       <span v-if="!mobileMinify && labelText" class="icon-wrapper-text" @click="handleScrollToComment">
-        <Icon :size="20" :icon="commentConfig?.icon">
-          <svg data-v-f0aeb853="" xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 1024 1024"><path fill="currentColor" d="M736 504a56 56 0 1 1 0-112 56 56 0 0 1 0 112m-224 0a56 56 0 1 1 0-112 56 56 0 0 1 0 112m-224 0a56 56 0 1 1 0-112 56 56 0 0 1 0 112M128 128v640h192v160l224-160h352V128z" /></svg>
+        <Icon :size="iconSize" :icon="commentConfig?.icon">
+          <svg data-v-f0aeb853="" xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 1024 1024"><path fill="currentColor" d="M736 504a56 56 0 1 1 0-112 56 56 0 0 1 0 112m-224 0a56 56 0 1 1 0-112 56 56 0 0 1 0 112m-224 0a56 56 0 1 1 0-112 56 56 0 0 1 0 112M128 128v640h192v160l224-160h352V128z" /></svg>
         </Icon>
         <span class="text">
           {{ labelText }}
         </span>
       </span>
       <span v-else class="icon-wrapper" @click="handleScrollToComment">
-        <Icon :size="20" :icon="commentConfig?.icon">
-          <svg data-v-f0aeb853="" xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 1024 1024"><path fill="currentColor" d="M736 504a56 56 0 1 1 0-112 56 56 0 0 1 0 112m-224 0a56 56 0 1 1 0-112 56 56 0 0 1 0 112m-224 0a56 56 0 1 1 0-112 56 56 0 0 1 0 112M128 128v640h192v160l224-160h352V128z" /></svg>
+        <Icon :size="iconSize" :icon="commentConfig?.icon">
+          <svg data-v-f0aeb853="" xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="0 0 1024 1024"><path fill="currentColor" d="M736 504a56 56 0 1 1 0-112 56 56 0 0 1 0 112m-224 0a56 56 0 1 1 0-112 56 56 0 0 1 0 112m-224 0a56 56 0 1 1 0-112 56 56 0 0 1 0 112M128 128v640h192v160l224-160h352V128z" /></svg>
         </Icon>
       </span>
     </div>
@@ -116,7 +122,7 @@ const showCommentBtn = computed(() => {
   position: fixed;
   width: v-bind(docWidth);
   text-align: right;
-  bottom: 40px;
+  bottom: v-bind(marginBottom);
   font-size: 16px;
   opacity: 0.6;
   display: flex;
@@ -176,7 +182,7 @@ html.dark .comment-btn-wrapper {
   transition-duration: 0.3s;
 }
 
-@supports (color: rgb(from red r g b / 1)) {
+@supports (backdrop-filter: blur(3px)) {
   .comment-btn-wrapper {
     opacity: 1;
     --box-shadow-color: rgba(0, 0, 0, 0.06);
@@ -186,10 +192,6 @@ html.dark .comment-btn-wrapper {
   }
   .comment-btn-wrapper :is(.icon-wrapper, .icon-wrapper-text) {
     backdrop-filter: blur(3px);
-  }
-  .comment-btn-wrapper :is(.icon-wrapper, .icon-wrapper-text):not(:hover):not(:active) {
-    background-color: rgb(from var(--vp-c-brand-soft) r g b / 0.6);
-    color: rgb(from var(--vp-c-brand-1) r g b / 0.6);
   }
 }